sarabotai/requirements.txt
# Requirements File: sarabotai/requirements.txt

streamlit==1.32.0
google-generativeai==0.3.2
langchain==0.1.5
chromadb==0.4.15
sentence-transformers==2.2.2
beautifulsoup4==4.12.2
requests==2.31.0
plotly==5.18.0
python-dotenv==1.0.0
pandas==2.0.3
unstructured==0.10.30
selenium==4.15.0
